components folder better organized in (a) User and (b) Admin folders and respective subfolders
