+++ components/Admin/Details/DetailsPlate/DetailsPlate.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import styles from "./DetailsPlate.module.css";
-import BackgroundCanvas from "../Shared/BackgroundCanvas/BackgroundCanvas";
-import TextParagraph from "../Shared/Typography/TextParagraph";
-import HashItem from "../hashItem/HashItem";
-import TextBold from "../Shared/Typography/TextBold/TextBold";
+import BackgroundCanvas from "../../Shared/BackgroundCanvas/BackgroundCanvas";
+import TextParagraph from "../../Shared/Typography/TextParagraph";
+import HashItem from "../../List/HashItem/HashItem";
+import TextBold from "../../Shared/Typography/TextBold/TextBold";
 import { dateFormat } from "@/utils/serialize";
 
 export default function DetailsPlate({
