+++ components/Admin/Details/StatusCard/StatusCard.module.css
