+++ components/Admin/List/IssueCard/IssueCard.jsx
@@ -1,12 +1,12 @@
 import React, { useState } from "react";
 import styles from "./IssueCard.module.css";
-import arrowDown from "../../../pages/assets/images/icon-arrow-down.svg";
+import arrowDown from "../../../../pages/assets/images/icon-arrow-down.svg";
 import Link from "next/link";
-import TextParagraph from "../Shared/Typography/TextParagraph";
-import HashItem from "../hashItem/HashItem";
+import TextParagraph from "../../Shared/Typography/TextParagraph";
+import HashItem from "../HashItem/HashItem";
 import { dateFormat } from "@/utils/serialize";
-import TextBold from "../Shared/Typography/TextBold";
-import BackgroundCanvas from "../Shared/BackgroundCanvas/BackgroundCanvas";
+import TextBold from "../../Shared/Typography/TextBold";
+import BackgroundCanvas from "../../Shared/BackgroundCanvas/BackgroundCanvas";
 
 export default function IssueCard(props) {
   const [showDescription, setShowDescription] = useState(false);
