+++ components/Admin/Nav/Avatar/Avatar.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import imgAvatar from "../../../pages/assets/images/image-avatar.jpg";
+import imgAvatar from "../../../../pages/assets/images/image-avatar.jpg";
 import styles from "./Avatar.module.css";
 
 export default function Avatar() {
