--- components/User/IssueForm/FormFieldLabel/FormFieldLabel.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import styles from "./FormFieldLabel.module.css";
-
-export default function FormFieldLabel() {
-  return <p>Name</p>;
-}
