+++ pages/admin/issues/[issueId].jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import prisma from "@/prisma/client";
 import serialize from "@/utils/serialize";
-import AdminTopBar from "@/components/Admin/AdminTopBar/AdminTopBar";
+import AdminTopBar from "@/components/Admin/Nav/AdminTopBar/AdminTopBar";
 import styles from "./IssueStatus.module.css";
 import arrowLeft from "../../../pages/assets/images/icon-arrow-left.svg";
-import IssueCard from "@/components/Admin/IssueCard/IssueCard";
-import StatusCard from "@/components/Admin/Shared/StatusCard/StatusCard";
-import DetailsPlate from "@/components/Admin/DetailsPlate/DetailsPlate";
+import IssueCard from "@/components/Admin/List/IssueCard/IssueCard";
+import StatusCard from "@/components/Admin/Details/StatusCard/StatusCard";
+import DetailsPlate from "@/components/Admin/Details/DetailsPlate/DetailsPlate";
 import Link from "next/link";
 import TextBold from "@/components/Admin/Shared/Typography/TextBold/TextBold";
